FF-2465 split out "Assignment Details" functionality into its own file
